refactor: update imports and remove deprecated types
- Refactored imports in various components to use the new models directory for enums and types.
- Removed obsolete type files related to base, icon, status, and variant, streamlining the codebase.
- Updated component logic to reflect changes in type imports, ensuring consistency across the application.
