+++ src/resources/assets/icons/Trash.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "@/shared/types/icon";
+import { IconProps } from "@/shared/models/icon";
 
 export function Trash({
   fill = '#fff',
