+++ src/shared/models/enums/operation.ts
@@ -1,0 +1,6 @@
+export enum Operation {
+    CREATING = 'creating',
+    UPDATING = 'updating',
+    DELETING = 'deleting',
+    VIEWING = 'viewing'
+}
