--- src/shared/types/base.ts
@@ -1,15 +1,0 @@
-  export enum Operation {
-    CREATING = 'creating',
-    UPDATING = 'updating',
-    DELETING = 'deleting',
-    VIEWING = 'viewing'
-  }
-  
-  export interface ElementConfig {
-    label: string;
-    error: string;
-    fullWidth: boolean;
-    className: string;
-    containerClassName: string;
-    options: Array<{ value: string; label: string }>;
-  }
