--- src/shared/types/icon.ts
@@ -1,5 +1,0 @@
-export type IconProps = {
-  fill?: string;
-  stroke?: string;
-  className?: string;
-};
