--- src/shared/types/variant.ts
@@ -1,6 +1,0 @@
-export enum Variant {
-    SUCCESS = 'success',
-    WARNING = 'warning',
-    DANGER = 'danger',
-    PRIMARY = 'primary'
-}
