+++ src/shared/ui/Button/Button.types.ts
@@ -1,11 +1,10 @@
 import { ButtonHTMLAttributes, ReactNode } from "react";
 
-export type ButtonVariant = 'primary' | 'secondary' | 'danger' | 'success' | 'outline' | 'warning';
-export type ButtonSize = 'sm' | 'md' | 'lg';
+import { Size, Variant } from "@/shared/models/types/base";
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  variant?: ButtonVariant;
-  size?: ButtonSize;
+  variant?: Variant;
+  size?: Size;
   isLoading?: boolean;
   fullWidth?: boolean;
   className?: string;
