+++ src/shared/ui/Card/Form/FormCard.types.ts
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
 
 import { Form } from "@/features/FormManagement/api/models/FormApiModel";
-import { Variant } from "@/shared/types/variant";
+import { Variant } from "@/shared/models/types/base";
 
 export type BadgeVariant = Variant;
