+++ src/shared/ui/Input/Input.type.ts
@@ -1,5 +1,5 @@
 import { InputHTMLAttributes } from 'react';
 
-import { ElementConfig } from '@/shared/types/base';
+import { ElementConfig } from '@/shared/models/types/base';
 
 export type InputProps = InputHTMLAttributes<HTMLInputElement> & Partial<ElementConfig>;
