+++ src/shared/ui/Select/Select.types.ts
@@ -1,5 +1,5 @@
 import { SelectHTMLAttributes } from 'react';
 
-import { ElementConfig } from '@/shared/types/base';
+import { ElementConfig } from '@/shared/models/types/base';
 
 export type SelectProps = SelectHTMLAttributes<HTMLSelectElement> & Partial<ElementConfig>;
